dfp_memryx/requirements.txt
absl-py==2.3.0
astunparse==1.6.3
attrs==21.4.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
-e git+https://github.com/serengil/deepface.git@efc00b1f88a5a38f4dbbffb162a7a51067885ede#egg=deepface
dlib==20.0.0
faiss-cpu==1.11.0
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==20181003210633
fonttools==4.58.4
fsspec==2025.5.1
gast==0.4.0
gdown==5.2.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
humanfriendly==10.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kagglehub==0.3.12
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
memryx==1.2.0
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.2
onnxruntime==1.18.1
onnxsim==0.4.36
opencv-contrib-python==4.8.1.78
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
posix-ipc==1.1.1
protobuf==3.20.3
psutil==6.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyNaCl==1.5.0
pyparsing==3.2.3
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qt-material==2.17
QtGraphVisuals==0.0.8
qtoml==0.3.1
requests==2.32.4
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rich-argparse==1.5.2
rsa==4.9.1
scipy==1.13.1
shiboken6==6.8.0.2
six==1.17.0
soupsieve==2.7
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-cpu-aws==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf2onnx==1.16.1
tf_keras==2.17.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
